+++ src/main/java/frc/robot/constants/TelemetryConstants.java
@@ -4,7 +4,7 @@ import edu.wpi.first.networktables.BooleanEntry;
 import edu.wpi.first.networktables.NetworkTable;
 import edu.wpi.first.networktables.NetworkTableInstance;
 import edu.wpi.first.wpilibj.smartdashboard.Field2d;
-import frc.robot.utils.Limelight;
+import frc.robot.limelight.Limelight;
 import frc.robot.utils.ShuffleboardTable;
 
 public class TelemetryConstants {
