+++ src/main/java/frc/robot/limelight/Limelight.java
@@ -1,4 +1,4 @@
-package frc.robot.utils;
+package frc.robot.limelight;
 
 import java.util.HashMap;
 import java.util.Map;
